# passIt-backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.SignupRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.User:
    properties:
      address:
        type: string
      created_at:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        description: User represents a user in the system
        type: string
      is_active:
        type: boolean
      is_admin:
        type: boolean
      keycloack_id:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  server.CreateUserRequestBody:
    properties:
      email:
        type: string
      first_name:
        type: string
      is_admin:
        type: boolean
      last_name:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  server.PassItResponseBody:
    properties:
      code:
        type: integer
      data: {}
    type: object
  server.UpdateUserRequestBody:
    properties:
      email:
        type: string
      first_name:
        type: string
      is_admin:
        description: Pointer to distinguish between false and not provided
        type: boolean
      last_name:
        type: string
      password:
        description: Optional password
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@passit.com
    name: API Support
  description: Event ticketing platform API with user management and authentication
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: PassIt API
  version: "1.0"
paths:
  /api/users:
    get:
      description: Retrieve a list of all users in the system
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all users (Admin only)
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with username, email, password and admin status
      parameters:
      - description: User creation data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/server.CreateUserRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.PassItResponseBody'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new user (Admin only)
      tags:
      - users
  /api/users/{id}:
    put:
      consumes:
      - application/json
      description: Update user information including email, name, password, and admin
        status
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User update data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/server.UpdateUserRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user by ID (Admin only)
      tags:
      - users
  /api/users/me:
    get:
      description: Retrieve the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /auth/callback:
    get:
      description: Handles the OAuth2 callback from Keycloak after authentication
      parameters:
      - description: Authorization code
        in: query
        name: code
        required: true
        type: string
      - description: State parameter
        in: query
        name: state
        required: true
        type: string
      responses:
        "302":
          description: Redirect to frontend
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: OAuth2 callback
      tags:
      - auth
  /auth/login:
    get:
      description: Redirects to Keycloak for authentication
      responses:
        "302":
          description: Redirect to Keycloak
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Initiate OAuth2 login
      tags:
      - auth
  /auth/logout:
    get:
      description: Logs the user out by deleting the session and clearing the cookie
      responses:
        "302":
          description: Redirect to Keycloak logout
          schema:
            type: string
      summary: Logout user
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Allows public users to self-register (always as non-admin)
      parameters:
      - description: User signup data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.SignupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register new user
      tags:
      - auth
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"